icon CUD links, fix confirm-delete, minor layout/formatting
Icon Hyperlinks:
- New supplyitem (on Nav)
- Update supplyitem
- Delete supplyitem

Models:
- get_url reverse functions for links

Bugfix:
- confirm/cancel delete form (previously both buttons would delete)

Item Template:
- Formatting warning levels - text fix show number 0 if no stock in
color white
- Links nice responsive spread apart layout

+++ chefwebsite/supplyitems/views.py
@@ -113,8 +113,11 @@ def delete(request, slug):
         obj = get_object_or_404(SupplyItem, slug=slug)
 
         if request.method == "POST":
-            obj.delete()
-            return redirect("list_supplyitems")
+            if request.POST.get("cancel"):
+                return redirect('list_supplyitems')
+            elif request.POST.get("confirm"):
+                obj.delete()
+                return redirect("list_supplyitems")
 
         context["object"] = obj
